python/requirements.txt
# Core ML dependencies (Apple Silicon compatible)
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.50.0
accelerate>=1.10.0
safetensors>=0.5.0
Pillow>=11.0.0
numpy>=2.0.0
opencv-python>=4.10.0
scipy>=1.14.0
scikit-image>=0.24.0
omegaconf>=2.3.0
timm>=1.0.0
einops>=0.8.0
huggingface_hub>=0.30.0

# Optimization dependencies
bitsandbytes>=0.45.0; sys_platform != "darwin"

# Watermark detection
scikit-learn>=1.6.0

# Benchmark metrics
lpips>=0.1.4
